requirements.txt
discord.py
python-dotenv
aiohttp
pynacl
aiosqlite
aiohttp
pillow
pilmoji
emoji==1.7.0